Modify: update.sh
install.shと同じく、セットアップスクリプト例外として外した。

+++ update.sh
@@ -6,6 +6,7 @@ git pull
 for f in .??*
 do
     [[ "$f" == ".git" ]] && continue
+    [[ "$f" == "*.sh" ]] && continue
     sudo rm -r "$HOME"/"$f"
     ln -snfv "$DOPATH"/"$f" "$HOME"/"$f"
 done
